requirements.txt
mecab-python3>=1.0.6
markovify>=0.9.4
Flask>=2.3.0
unidic>=1.0.0
Misskey.py>=0.4.0
Mastodon.py>=1.8.0
python-Levenshtein>=0.21.0
sentry-sdk[flask]>=1.32.0
gunicorn>=21.0.0
requests>=2.31.0